+++ CriticalMass/DataStore.swift
@@ -7,8 +7,10 @@
 
 import Foundation
 
+public typealias ApiResponseResult = Result<ApiResponse, NetworkError>
+
 public protocol DataStore {
-    func update(with response: ApiResponse)
+    func update(with response: ApiResponseResult)
 
     func add(friend: Friend)
     func remove(friend: Friend)