+++ CriticalMass/Notification+Static.swift
@@ -15,4 +15,5 @@ extension Notification.Name {
     static let positionOthersChanged = Notification.Name(rawValue: "positionOthersChanged")
     static let chatMessagesReceived = Notification.Name(rawValue: "chatMessagesReceived")
     static let focusLocation = Notification.Name("focusLocation")
+    static let locationAndMessagesReceived = Notification.Name("locationAndMessagesReceived")
 }